+++ taf/tests/test_yubikey/conftest.py
@@ -3,7 +3,7 @@ from taf.tests import TEST_WITH_REAL_YK
 from taf.tests.conftest import KEYSTORE_PATH
 
 from pytest import fixture
-from taf.tests.yubikey_utils import TargetYubiKey, _yk_piv_ctrl_mock
+from taf.tools.yubikey.yubikey_utils import TargetYubiKey, _yk_piv_ctrl_mock
 
 
 def pytest_configure(config):
